+++ src/App.jsx
@@ -6,7 +6,6 @@ import Home from './components/Home';
 import About from './components/About';
 import Photos from './components/Photos';
 import Contacts from './components/Contacts';
-import Frame from './components/Frame';
 
 
 const App = () => {
@@ -19,7 +18,6 @@ const App = () => {
         <Route path="/photos" element={<Photos />} /> {/* Use 'element' prop */}
         <Route path="/contacts" element={<Contacts />} /> {/* Use 'element' prop */}
         <Route path="/" element={<Home />} /> {/* Use 'element' prop */}
-        <Route path="/frame" element={<Frame />} />
       </Routes>
     </div>
   </Router>
