+++ src/components/NavBar.jsx
@@ -15,19 +15,16 @@ const NavBar = () => {
         </div>
       <ul className="nav-list"> {/* Use className instead of style */}
         <li className="nav-item"> {/* Use className instead of style */}
-          <Link to="/">Home</Link>
+          <Link to="/">HOME</Link>
         </li>
         <li className="nav-item"> 
-          <Link to="/about">About</Link>
+          <Link to="/about">ABOUT</Link>
         </li>
         <li className="nav-item"> 
-          <Link to="/photos">Photos</Link>
+          <Link to="/photos">PHOTOS</Link>
         </li>
         <li className="nav-item"> 
-          <Link to="/contacts">Contacts</Link>
-        </li>
-        <li className="nav-item"> 
-          <Link to="/frame">Frame</Link>
+          <Link to="/contacts">CONTACT</Link>
         </li>
       </ul>
     </nav>
