+++ src/styles/Home.css
@@ -44,32 +44,33 @@ html, body {
   .frame .rectangle {
     background-color: #ffffff;
     height: 87px;
-    left: 18px;
-    position: absolute;
-    top: 993px;
-    width: 1107px;
+    left: -16%;
+    position: relative;
+    top: 880px;
+    width: 120%;
   }
   
   .frame .div {
     background-color: #8c5581;
-    height: 985px;
-    left: 1115px;
+    height: 990px;
+    left: 96%;
     position: absolute;
-    top: 95px;
+    top: 9px;
     width: 343px;
   }
   
   .frame .circle-edit {
-    background-image: url(./ellipse-5.png);
+    background-image: url(src/photos/pexels-photo-2562555.jpeg);
     background-position: 50% 50%;
     background-size: cover;
-    height: 606px;
-    left: 738px;
+    height: 580px; /* Adjusted height to make it a square */
+    width: 580px;
+    border-radius: 50%; /* Make it a circle */
     position: absolute;
-    top: 284px;
-    width: 591px;
-  }
-  
+    left: 1000px;
+    top: 210px;
+}
+
   .frame .navbar {
       display: flex;
       align-items: center;
@@ -93,9 +94,9 @@ html, body {
   .frame .rectangle-2 {
     background-color: #8c5581;
     height: 220px;
-    left: 0;
+    right: 100%;
     position: absolute;
-    top: 809px;
+    top: 700px;
     width: 297px;
   }
   
@@ -164,14 +165,14 @@ html, body {
   .frame .welcome-to-the {
     color: #000000;
     font-family: "Lexend Deca-Regular", Helvetica;
-    font-size: 40px;
+    font-size: 50px;
     font-weight: 400;
-    left: 0;
+    right: 13%;
     letter-spacing: 4.8px;
     line-height: normal;
     position: absolute;
     top: 0;
-    width: 514px;
+    width: 600px;
   }
   
   .frame .text-wrapper-5 {
@@ -179,7 +180,7 @@ html, body {
     font-family: "Lexend Deca-Regular", Helvetica;
     font-size: 20px;
     font-weight: 400;
-    left: 0;
+    right: 52%;
     letter-spacing: 2.4px;
     line-height: normal;
     position: absolute;
@@ -193,12 +194,12 @@ html, body {
     font-family: "Lexend Deca-Regular", Helvetica;
     font-size: 50px;
     font-weight: 400;
-    left: 327px;
+    right: 30%;
     letter-spacing: 6px;
     line-height: normal;
-    position: absolute;
+    position: relative;
     text-decoration: underline;
-    top: 788px;
+    top: 400px;
     width: 196px;
   }
   
@@ -207,11 +208,11 @@ html, body {
     font-family: "Lexend Deca-Regular", Helvetica;
     font-size: 20px;
     font-weight: 400;
-    left: 327px;
+    right: 30%;
     letter-spacing: 2.4px;
     line-height: normal;
-    position: absolute;
-    top: 869px;
+    position: relative;
+    top: 410px;
     width: 180px;
   }
